refactor(app): wave-15 Slice D test 400-line ratchet and densify (#4352)
Ratchet repo.app_test_file_size 500→400 with shrink-only allowlist for
remaining offenders; extract shared fixtures for shortcut-host twins,
production labour, e2e barrel, region-map part3, and app event bus pins.
