+++ app/test/combat_ui_specs_part1_test.dart
@@ -16,7 +16,6 @@ import 'package:colonizethis_app/features/game/widgets/combat/quick_battle_deplo
 import 'package:colonizethis_app/widgets/ct_dialog_shell.dart';
 import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
 
-
 import 'combat_ui_specs_test_support.dart';
 
 QuickBattleGroup _centerFront({
