+++ app/test/move_army_invasion_intel_test.dart
@@ -1,6 +1,5 @@
 // Pins move army invasion intel helper and DLG20001 invasion rows (#4216).
 
-
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
