+++ app/test/province_build_port_shortcut_host_emit_event_test.dart
@@ -1,30 +1,17 @@
 // Pins the host-level Build port *shortcut-assignment* tap flow for both
-// province detail hosts (`GameMapProvinceDetailSidePanel` wide,
-// `GameMapNarrowDetailOverlaySlot` narrow). Refs #4332.
+// province detail hosts. Refs #4332, #4352.
 
 import 'package:colonizethis_app/config/constants.dart';
-import 'package:colonizethis_app/core/services/game_service/game_service.dart';
-import 'package:colonizethis_app/features/game/flame/overlays/game_map_narrow_detail_overlay.dart';
-import 'package:colonizethis_app/features/game/flame/overlays/game_map_province_detail_side_panel.dart';
-import 'package:colonizethis_app/features/game/flame/caches/per_player_work_target_selection_cache.dart';
-import 'package:colonizethis_app/providers/app_event_bus_provider.dart';
-import 'package:colonizethis_app/providers/game_service_provider.dart';
-import 'package:colonizethis_app/providers/games_box_provider.dart';
-import 'package:colonizethis_app/providers/games_provider.dart';
-import 'package:colonizethis_app/providers/map_province_panel_provider.dart';
 import 'package:colonizethis_app/widgets/ct_icon_action.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart' show buildPlayerView;
 import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_save/colonizethis_save.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
-import 'app_shell_harness.dart';
+import 'province_shortcut_host_emit_test_support.dart';
 
 const String _kGameId = 'g_bp_shortcut_emit';
 const String _kHumanPlayerId = 'gp1';
@@ -83,27 +70,6 @@ final Map<String, TileMapResult> _tileMapByRegion = {
     ],
   ),
 };
-
-class _GameServiceBuildPort extends GameService {
-  _GameServiceBuildPort(super.box, super.adapter);
-
-  @override
-  ({
-    MapTopology combinedTopology,
-    Map<String, TileMapResult> tileMapByRegion,
-    Map<String, MapTopology> topologyByRegion,
-    List<WarpLink>? warpLinks,
-  })?
-  getMapData(String gameId) {
-    if (gameId != _kGameId) return null;
-    return (
-      combinedTopology: _combinedTopology,
-      tileMapByRegion: _tileMapByRegion,
-      topologyByRegion: _topologyByRegion,
-      warpLinks: null,
-    );
-  }
-}
 
 Game _buildGame({required bool withEngineer}) {
   return Game(
@@ -187,21 +153,6 @@ RegionMapViewData _region() {
   );
 }
 
-PerPlayerWorkTargetSelectionCache _refreshedCache(Game game) {
-  final playerView = buildPlayerView(game, _combinedTopology, _kHumanPlayerId);
-  return PerPlayerWorkTargetSelectionCache()
-    ..refresh(
-      WorkTargetSelectionSnapshot(
-        game: game,
-        playerId: _kHumanPlayerId,
-        playerView: playerView,
-        topology: _combinedTopology,
-        currentOrders: const Orders(),
-        tileMapByRegion: _tileMapByRegion,
-      ),
-    );
-}
-
 Finder _buildPortAction({required bool enabledOnly}) {
   return find.byWidgetPredicate(
     (Widget w) =>
@@ -210,31 +161,6 @@ Finder _buildPortAction({required bool enabledOnly}) {
         (!enabledOnly || w.onPressed != null),
   );
 }
-
-typedef _HostCase = ({
-  String label,
-  Type hostType,
-  Size surfaceSize,
-  bool selectTileTab,
-  bool wide,
-});
-
-const List<_HostCase> _hostCases = <_HostCase>[
-  (
-    label: 'The wide side panel',
-    hostType: GameMapProvinceDetailSidePanel,
-    surfaceSize: Size(720, 720),
-    selectTileTab: false,
-    wide: true,
-  ),
-  (
-    label: 'The narrow bottom-slot host',
-    hostType: GameMapNarrowDetailOverlaySlot,
-    surfaceSize: Size(400, 600),
-    selectTileTab: true,
-    wide: false,
-  ),
-];
 
 void main() {
   suppressLogsForTests();
@@ -249,71 +175,31 @@ void main() {
   Future<List<OpenCivilianUnitsPanelEvent>> pumpHostAndSelect(
     WidgetTester tester, {
     required Game game,
-    required _HostCase host,
-  }) async {
-    final region = _region();
-    final playerView = buildPlayerView(game, _combinedTopology, _kHumanPlayerId);
-    final cache = _refreshedCache(game);
-    final Widget body = host.wide
-        ? Center(
-            child: SizedBox(
-              width: 320,
-              child: GameMapProvinceDetailSidePanel(
-                game: game,
-                region: region,
-                humanPlayerId: _kHumanPlayerId,
-                playerView: playerView,
-                workTargetSelectionCache: cache,
-              ),
-            ),
-          )
-        : Align(
-            alignment: Alignment.bottomCenter,
-            child: GameMapNarrowDetailOverlaySlot(
-              game: game,
-              region: region,
-              humanPlayerId: _kHumanPlayerId,
-              playerView: playerView,
-              workTargetSelectionCache: cache,
-            ),
-          );
-
-    final bus = AppEventBus.create();
-    addTearDown(bus.dispose);
-    final opened = <OpenCivilianUnitsPanelEvent>[];
-    final sub = bus.on<OpenCivilianUnitsPanelEvent>().listen(opened.add);
-    addTearDown(sub.cancel);
-
-    await pumpAppShell(
-      tester,
-      viewport: host.surfaceSize,
-      overrides: [
-        gamesBoxProvider.overrideWith((ref) => gamesBox),
-        gameServiceProvider.overrideWith(
-          (ref) => _GameServiceBuildPort(gamesBox, GameSaveAdapter()),
+    required ProvinceShortcutHostCase host,
+  }) =>
+      pumpProvinceShortcutHostAndSelect(
+        tester,
+        gamesBox: gamesBox,
+        gameService: provinceShortcutHostEmitGameService(
+          gamesBox: gamesBox,
+          gameId: _kGameId,
+          combinedTopology: _combinedTopology,
+          tileMapByRegion: _tileMapByRegion,
+          topologyByRegion: _topologyByRegion,
         ),
-        appEventBusProvider.overrideWith((ref) => bus),
-        currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-        currentOrdersProvider.overrideWith(
-          () => CurrentOrdersNotifier(const Orders()),
+        game: game,
+        humanPlayerId: _kHumanPlayerId,
+        host: host,
+        region: _region(),
+        combinedTopology: _combinedTopology,
+        workTargetSelectionCache: refreshedProvinceShortcutWorkTargetCache(
+          game: game,
+          humanPlayerId: _kHumanPlayerId,
+          combinedTopology: _combinedTopology,
+          tileMapByRegion: _tileMapByRegion,
         ),
-      ],
-      child: Scaffold(body: body),
-    );
-
-    final ctx = tester.element(find.byType(host.hostType));
-    ProviderScope.containerOf(ctx)
-        .read(mapProvincePanelProvider.notifier)
-        .reportMapTileTapped(_kTileKey);
-    await tester.pumpAndSettle();
-    if (host.selectTileTab) {
-      final tileTab = find.text('Tile');
-      expect(tileTab, findsOneWidget);
-      await tester.tap(tileTab);
-      await tester.pumpAndSettle();
-    }
-    return opened;
-  }
+        selectedTileKey: _kTileKey,
+      );
 
   Future<void> expectBuildPortShortcutEmits(
     WidgetTester tester, {
@@ -321,40 +207,22 @@ void main() {
     required String hostLabel,
   }) async {
     final shortcut = _buildPortAction(enabledOnly: true);
-    expect(
-      shortcut,
-      findsOneWidget,
-      reason:
-          '$hostLabel must render an enabled Build port inline action '
-          'for a valid Engineer + affordable road tile.',
-    );
+    expect(shortcut, findsOneWidget, reason: '$hostLabel enabled Build port');
     await tester.ensureVisible(shortcut);
     await tester.tap(shortcut);
     await tester.pump();
-    expect(
-      opened,
-      hasLength(1),
-      reason:
-          'Tapping the enabled Build port shortcut must open the '
-          'Civilian Units panel exactly once via OpenCivilianUnitsPanelEvent.',
-    );
+    expect(opened, hasLength(1));
     final event = opened.single;
     expect(event.engineerOnly, isTrue);
     expect(event.builderOnly, isFalse);
     expect(event.explorerOnly, isFalse);
-    expect(
-      event.buildPortShortcutTargetTileKey,
-      _kTileKey,
-      reason:
-          'The shortcut must target the exact selected tile key so the '
-          'Engineer-only panel can assign build_port to that tile.',
-    );
+    expect(event.buildPortShortcutTargetTileKey, _kTileKey);
     expect(event.exploreShortcutTargetTileKey, isNull);
     expect(event.prospectShortcutTargetTileKey, isNull);
     expect(event.buildImprovementShortcutTargetTileKey, isNull);
   }
 
-  for (final host in _hostCases) {
+  for (final host in provinceShortcutHostCases) {
     testWidgets(
       '${host.wide ? 'wide' : 'narrow'} host: tapping the enabled Build '
       'road shortcut emits an Engineer-only OpenCivilianUnitsPanelEvent '
@@ -397,13 +265,7 @@ void main() {
             await tester.tap(anyShortcut.first, warnIfMissed: false);
             await tester.pump();
           }
-          expect(
-            opened,
-            isEmpty,
-            reason:
-                'A disabled / absent Build port shortcut must never '
-                'open the Civilian Units panel via OpenCivilianUnitsPanelEvent.',
-          );
+          expect(opened, isEmpty);
         } else {
           expect(opened, isEmpty);
         }
