+++ app/test/province_upgrade_town_shortcut_host_emit_event_test.dart
@@ -1,46 +1,21 @@
 // Pins the host-level Upgrade town *shortcut-assignment* tap flow for
-// both province detail hosts (`GameMapProvinceDetailSidePanel` wide,
-// `GameMapNarrowDetailOverlaySlot` narrow).
-//
-// SPEC: SPEC/ui/province-sea-zone-detail-overlay.md
-// § Political Upgrade town shortcut behavior:
-//   On tap, open Civilian Units panel in Builder-only shortcut mode targeting
-//   upgrade_town for the province town tile key.
-// SPEC/ui/civilian-units-panel.md — explicit shortcut contract:
-//   `upgradeTownShortcutTargetTileKey` opens direct-assign `upgrade_town`.
-//
-// Coverage gap closed here (Refs #4316):
-//   - `province_overlay_political_upgrade_town_test.dart` pins gist helper copy.
-//   - This file pins the *host wiring*: that tapping the enabled Political
-//     Upgrade town control emits `OpenCivilianUnitsPanelEvent(builderOnly: true,
-//     upgradeTownShortcutTargetTileKey: <province town tile key>)` on the app
-//     event bus, plus the no-Builder negative.
+// both province detail hosts. SPEC/ui/province-sea-zone-detail-overlay.md;
+// SPEC/ui/civilian-units-panel.md. Refs #4316, #4352.
 
 import 'package:colonizethis_app/config/constants.dart';
-import 'package:colonizethis_app/core/services/game_service/game_service.dart';
 import 'package:colonizethis_app/features/game/flame/map_state/game_map_area_state_logic.dart';
-import 'package:colonizethis_app/features/game/flame/overlays/game_map_narrow_detail_overlay.dart';
-import 'package:colonizethis_app/features/game/flame/overlays/game_map_province_detail_side_panel.dart';
-import 'package:colonizethis_app/features/game/flame/caches/per_player_work_target_selection_cache.dart';
-import 'package:colonizethis_app/providers/app_event_bus_provider.dart';
-import 'package:colonizethis_app/providers/game_service_provider.dart';
-import 'package:colonizethis_app/providers/games_box_provider.dart';
-import 'package:colonizethis_app/providers/games_provider.dart';
-import 'package:colonizethis_app/providers/map_province_panel_provider.dart';
 import 'package:colonizethis_app/widgets/ct_action_text_button.dart';
-import 'package:colonizethis_app_l10n/l10n/app_localizations_en.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart' show buildPlayerView;
 import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_save/colonizethis_save.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
+import 'package:colonizethis_app_l10n/l10n/app_localizations_en.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
-import 'app_shell_harness.dart';
+import 'province_shortcut_host_emit_test_support.dart';
 
 const String _kGameId = 'g_ut_shortcut_emit';
 const String _kHumanPlayerId = 'gp1';
@@ -96,27 +71,6 @@ final Map<String, TileMapResult> _tileMapByRegion = {
     ],
   ),
 };
-
-class _GameServiceUpgradeTownShortcut extends GameService {
-  _GameServiceUpgradeTownShortcut(super.box, super.adapter);
-
-  @override
-  ({
-    MapTopology combinedTopology,
-    Map<String, TileMapResult> tileMapByRegion,
-    Map<String, MapTopology> topologyByRegion,
-    List<WarpLink>? warpLinks,
-  })?
-  getMapData(String gameId) {
-    if (gameId != _kGameId) return null;
-    return (
-      combinedTopology: _combinedTopology,
-      tileMapByRegion: _tileMapByRegion,
-      topologyByRegion: _topologyByRegion,
-      warpLinks: null,
-    );
-  }
-}
 
 Game _buildGame({required bool withBuilder}) {
   return Game(
@@ -201,21 +155,6 @@ RegionMapViewData _region() {
   );
 }
 
-PerPlayerWorkTargetSelectionCache _refreshedCache(Game game) {
-  final playerView = buildPlayerView(game, _combinedTopology, _kHumanPlayerId);
-  return PerPlayerWorkTargetSelectionCache()
-    ..refresh(
-      WorkTargetSelectionSnapshot(
-        game: game,
-        playerId: _kHumanPlayerId,
-        playerView: playerView,
-        topology: _combinedTopology,
-        currentOrders: const Orders(),
-        tileMapByRegion: _tileMapByRegion,
-      ),
-    );
-}
-
 Finder _upgradeTownAction({required bool enabledOnly}) {
   final l10n = AppLocalizationsEn();
   return find.byWidgetPredicate(
@@ -226,27 +165,7 @@ Finder _upgradeTownAction({required bool enabledOnly}) {
   );
 }
 
-typedef _HostCase = ({
-  String label,
-  Type hostType,
-  Size surfaceSize,
-  bool wide,
-});
-
-const List<_HostCase> _hostCases = <_HostCase>[
-  (
-    label: 'The wide side panel',
-    hostType: GameMapProvinceDetailSidePanel,
-    surfaceSize: Size(720, 720),
-    wide: true,
-  ),
-  (
-    label: 'The narrow bottom-slot host',
-    hostType: GameMapNarrowDetailOverlaySlot,
-    surfaceSize: Size(400, 600),
-    wide: false,
-  ),
-];
+const List<ProvinceShortcutHostCase> _hostCases = provinceShortcutHostCases;
 
 void main() {
   suppressLogsForTests();
@@ -278,65 +197,31 @@ void main() {
   Future<List<OpenCivilianUnitsPanelEvent>> pumpHostAndSelect(
     WidgetTester tester, {
     required Game game,
-    required _HostCase host,
-  }) async {
-    final region = _region();
-    final playerView = buildPlayerView(game, _combinedTopology, _kHumanPlayerId);
-    final cache = _refreshedCache(game);
-    final Widget body = host.wide
-        ? Center(
-            child: SizedBox(
-              width: 320,
-              child: GameMapProvinceDetailSidePanel(
-                game: game,
-                region: region,
-                humanPlayerId: _kHumanPlayerId,
-                playerView: playerView,
-                workTargetSelectionCache: cache,
-              ),
-            ),
-          )
-        : Align(
-            alignment: Alignment.bottomCenter,
-            child: GameMapNarrowDetailOverlaySlot(
-              game: game,
-              region: region,
-              humanPlayerId: _kHumanPlayerId,
-              playerView: playerView,
-              workTargetSelectionCache: cache,
-            ),
-          );
-
-    final bus = AppEventBus.create();
-    addTearDown(bus.dispose);
-    final opened = <OpenCivilianUnitsPanelEvent>[];
-    final sub = bus.on<OpenCivilianUnitsPanelEvent>().listen(opened.add);
-    addTearDown(sub.cancel);
-
-    await pumpAppShell(
-      tester,
-      viewport: host.surfaceSize,
-      overrides: [
-        gamesBoxProvider.overrideWith((ref) => gamesBox),
-        gameServiceProvider.overrideWith(
-          (ref) => _GameServiceUpgradeTownShortcut(gamesBox, GameSaveAdapter()),
+    required ProvinceShortcutHostCase host,
+  }) =>
+      pumpProvinceShortcutHostAndSelect(
+        tester,
+        gamesBox: gamesBox,
+        gameService: provinceShortcutHostEmitGameService(
+          gamesBox: gamesBox,
+          gameId: _kGameId,
+          combinedTopology: _combinedTopology,
+          tileMapByRegion: _tileMapByRegion,
+          topologyByRegion: _topologyByRegion,
         ),
-        appEventBusProvider.overrideWith((ref) => bus),
-        currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-        currentOrdersProvider.overrideWith(
-          () => CurrentOrdersNotifier(const Orders()),
+        game: game,
+        humanPlayerId: _kHumanPlayerId,
+        host: provinceShortcutHostCaseWithoutTileTab(host),
+        region: _region(),
+        combinedTopology: _combinedTopology,
+        workTargetSelectionCache: refreshedProvinceShortcutWorkTargetCache(
+          game: game,
+          humanPlayerId: _kHumanPlayerId,
+          combinedTopology: _combinedTopology,
+          tileMapByRegion: _tileMapByRegion,
         ),
-      ],
-      child: Scaffold(body: body),
-    );
-
-    final ctx = tester.element(find.byType(host.hostType));
-    ProviderScope.containerOf(ctx)
-        .read(mapProvincePanelProvider.notifier)
-        .reportMapTileTapped(_kTileKey);
-    await tester.pumpAndSettle();
-    return opened;
-  }
+        selectedTileKey: _kTileKey,
+      );
 
   Future<void> expectUpgradeTownShortcutEmits(
     WidgetTester tester, {
@@ -344,23 +229,11 @@ void main() {
     required String hostLabel,
   }) async {
     final shortcut = _upgradeTownAction(enabledOnly: true);
-    expect(
-      shortcut,
-      findsOneWidget,
-      reason:
-          '$hostLabel must render an enabled Upgrade town political control for '
-          'a valid Builder + National Bureaucracy + upgradeable town.',
-    );
+    expect(shortcut, findsOneWidget, reason: '$hostLabel enabled Upgrade town');
     await tester.ensureVisible(shortcut);
     await tester.tap(shortcut);
     await tester.pump();
-    expect(
-      opened,
-      hasLength(1),
-      reason:
-          'Tapping the enabled Upgrade town shortcut must open the Civilian '
-          'Units panel exactly once via OpenCivilianUnitsPanelEvent.',
-    );
+    expect(opened, hasLength(1));
     final event = opened.single;
     expect(event.builderOnly, isTrue);
     expect(event.explorerOnly, isFalse);
@@ -379,8 +252,7 @@ void main() {
     testWidgets(
       '${host.wide ? 'wide' : 'narrow'} host: tapping the enabled Upgrade '
       'town shortcut emits a Builder-only OpenCivilianUnitsPanelEvent '
-      'targeting the province town tile key (SPEC § Political Upgrade town '
-      'shortcut assignment)',
+      'targeting the province town tile key',
       (WidgetTester tester) async {
         final opened = await pumpHostAndSelect(
           tester,
