+++ test/check_app_test_file_size_test.dart
@@ -45,7 +45,7 @@ void main() {
     Directory('${temp.path}/app/test').createSync(recursive: true);
     File('${temp.path}/app/test/huge_test.dart')
       ..createSync()
-      ..writeAsStringSync(List.filled(501, '// line').join('\n'));
+      ..writeAsStringSync(List.filled(401, '// line').join('\n'));
 
     final logs = <String>[];
     final code = runCheckAppTestFileSize(
